requirements.txt
aiohttp==3.8.1
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.4.0
django-celery-results==2.3.1
django-filter==21.1
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
gunicorn==20.1.0
kombu==5.2.4
Markdown==3.3.7
packaging==21.3
pep517==0.12.0
pip-tools==6.6.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
pytest==7.3.1
pytest-html==3.2.0
pytest-cov==4.0.0
pytest-django==4.7.0
pytest-sugar==0.9.7